+++ packages/core-components/src/lib/unsorted/ui/Details.svelte
@@ -1,0 +1,16 @@
+<script context="module">
+	export const evidenceInclude = true;
+</script>
+
+<script>
+	export let title = 'Details';
+</script>
+
+<details class="mb-4 mt-3 ml-1">
+	<summary class="italic text-base text-grey-800 cursor-pointer">
+		{title}
+	</summary>
+	<div class="ml-3 pt-3 mb-6 text-base">
+		<slot />
+	</div>
+</details>