+++ packages/plugin-connector/src/plugin-discovery/load-config.spec.js
@@ -12,13 +12,11 @@ import {
 	validConfigParsed,
 	validMinimalConfig,
 	validMinimalConfigParsed
-} from './resolve-evidence-config.fixture';
-import { loadConfig } from './resolve-evidence-config';
+} from './load-config.fixture';
+import { loadConfig } from './load-config';
 
-/** @type {import("vitest").MockedFunction<typeof import("fs/promises").readFile>} */
-let mockedReadFile = /** @type {import("vitest").MockedFunction<typeof fs.readFile>} */ (
-	fs.readFile
-);
+fs.readFile = /** @type {any} */ (vi.fn());
+const mockedReadFile = vi.mocked(fs.readFile);
 
 describe('loadConfig', () => {
 	it('should load a valid configuration', async () => {