+++ packages/plugin-connector/src/plugin-discovery/resolve-evidence-config.js
@@ -1,55 +1,6 @@
-import chalk from 'chalk';
-import fs from 'fs/promises';
-import yaml from 'yaml';
-import { EvidenceConfigSchema } from './schemas/evidence-config.schema';
 import { isValidPackage } from './is-valid-package';
-import { cleanZodErrors } from '../lib/clean-zod-errors';
 import path from 'path';
-
-/**
- *
- * @param {string} rootDir
- * @returns {Promise<EvidenceConfig>}
- */
-export const loadConfig = async (rootDir) => {
-	const configPath = `${rootDir}/evidence.plugins.yaml`;
-	try {
-		const configFileContent = await fs.readFile(configPath, 'utf8').then((r) => r.toString());
-		// Surround all YAML key that begin with "@" in quotes
-		// Skipping keys that are already quoted (e.g. beginning of line or whitespace)
-		const rawConfig = yaml.parse(configFileContent.replaceAll(/($|\s)(@.+):/g, '$1"$2":'));
-
-		const configResult = EvidenceConfigSchema.safeParse(rawConfig);
-		if (!configResult.success) {
-			console.error(
-				chalk.bold.red(
-					`[!] evidence.plugins.yaml does not contain a valid configuration. \n    Plugins will not be loaded. This may lead to unexpected behavior.`
-				)
-			);
-			const formattedError = cleanZodErrors(configResult.error.format());
-			console.error(chalk.red('|   Discovered Errors:'));
-			const redPipe = chalk.red('|');
-			console.error(
-				`${redPipe}   ${yaml.stringify(formattedError).replace(/\n/g, `\n${redPipe}   `)}`
-			);
-			throw new Error('Invalid evidence.plugins.yaml');
-		}
-
-		return configResult.data;
-	} catch (e) {
-		if (!(e instanceof Error)) throw e;
-		if (e.message.startsWith('ENOENT')) {
-			console.warn('Could not find evidence plugins file. Using defaults.');
-			return EvidenceConfigSchema.parse({
-				components: {
-					'@evidence-dev/core-components': {}
-				}
-			});
-		}
-		throw e;
-	}
-};
-
+import { loadConfig } from './load-config';
 /**
  * Wrapper function to create a package validator function
  * @param {string} rootDir
@@ -79,7 +30,7 @@ const validatePlugin =
  */
 export const resolveEvidencePackages = async (rootDir) => {
 	/** @type {EvidenceConfig} */
-	const configContent = await loadConfig(rootDir);
+	const configContent = loadConfig(rootDir);
 
 	/** @type {EvidencePluginPackage<ValidPackage>[]} */
 	const componentPackages = await Promise.all(