+++ packages/sqlite/index.cjs
@@ -15,10 +15,9 @@ const envMap = {
 /** @type {import('@evidence-dev/db-commons').RunQuery<SQLiteOptions>} */
 const runQuery = async (queryString, database) => {
 	const filename = database ? database.filename : getEnv(envMap, 'filename');
-	const filepath = filename !== ':memory:' ? '../../' + filename : filename;
 	try {
 		const db = await open({
-			filename: filepath,
+			filename: filename,
 			driver: sqlite3.Database,
 			mode: sqlite3.OPEN_READONLY
 		});
@@ -45,10 +44,15 @@ module.exports = runQuery;
  */
 
 /** @type {import('@evidence-dev/db-commons').GetRunner<SQLiteOptions>} */
-module.exports.getRunner = async (opts) => {
+module.exports.getRunner = async (opts, directory) => {
+	/**
+	 * @param {string} queryContent
+	 * @param {string} queryPath
+	 * @returns {Promise<QueryResult>}
+	 */
 	return async (queryContent, queryPath) => {
 		// Filter out non-sql files
 		if (!queryPath.endsWith('.sql')) return null;
-		return runQuery(queryContent, opts);
+		return runQuery(queryContent, { ...opts, filename: path.join(directory, opts.filename) });
 	};
 };