+++ packages/universal-sql/src/cache-duckdb.d.ts
@@ -1,16 +1,14 @@
 /**
  * Caches DuckDB queries that have been
- * @param {string} route_hash md5 hash of the route id
  * @param {string} sql_string SQL string executed
- * @param {string} query_name name of the query
  * @param {import("apache-arrow").Table} data result of the query
+ * @param {{ route_hash: string, query_name: string, prerendering: boolean }} cache_options
  * @returns {void}
  */
 export function cache_for_hash(
-	route_hash: string,
 	sql_string: string,
-	query_name: string,
-	data: import('apache-arrow').Table
+	data: import('apache-arrow').Table,
+	cache_options: { route_hash: string; query_name: string; prerendering: boolean }
 ): void;
 
 /**