+++ packages/universal-sql/src/client-duckdb/node.js
@@ -1,4 +1,4 @@
-import { arrowTableToJSON } from './both.js';
+import { arrowTableToJSON, getPromise } from './both.js';
 import {
 	ConsoleLogger,
 	createDuckDB,
@@ -8,6 +8,7 @@ import {
 import { createRequire } from 'module';
 import { dirname, resolve } from 'path';
 import { cache_for_hash } from '../cache-duckdb.js';
+import { withTimeout } from './both.js';
 
 const require = createRequire(import.meta.url);
 const DUCKDB_DIST = dirname(require.resolve('@duckdb/duckdb-wasm'));
@@ -20,31 +21,49 @@ let db;
 /** @type {import("@duckdb/duckdb-wasm/dist/types/src/bindings/connection").DuckDBConnection} */
 let connection;
 
+const { resolve: resolveInit, reject: rejectInit, promise: initPromise } = getPromise();
+let initializing = false;
+
 /**
  * Initializes the database.
  *
  * @returns {Promise<void>}
  */
 export async function initDB() {
+	// If the database is already available, don't do anything
 	if (db) return;
 
-	const DUCKDB_BUNDLES = {
-		mvp: {
-			mainModule: resolve(DUCKDB_DIST, './duckdb-mvp.wasm'),
-			mainWorker: resolve(DUCKDB_DIST, './duckdb-node-mvp.worker.cjs')
-		},
-		eh: {
-			mainModule: resolve(DUCKDB_DIST, './duckdb-eh.wasm'),
-			mainWorker: resolve(DUCKDB_DIST, './duckdb-node-eh.worker.cjs')
-		}
-	};
-	const logger = new ConsoleLogger();
+	// If the database is already initializing, don't try to do it twice
+	// Instead, let the call wait for the initPromise
+	if (initializing) return withTimeout(initPromise);
 
-	// and synchronous database
-	db = await createDuckDB(DUCKDB_BUNDLES, logger, NODE_RUNTIME);
-	await db.instantiate();
-	db.open({ query: { castBigIntToDouble: true, castTimestampToDate: true } });
-	connection = db.connect();
+	// This call is the first (to execute), don't let anybody else try
+	// to initialize the database
+	initializing = true;
+
+	try {
+		const DUCKDB_BUNDLES = {
+			mvp: {
+				mainModule: resolve(DUCKDB_DIST, './duckdb-mvp.wasm'),
+				mainWorker: resolve(DUCKDB_DIST, './duckdb-node-mvp.worker.cjs')
+			},
+			eh: {
+				mainModule: resolve(DUCKDB_DIST, './duckdb-eh.wasm'),
+				mainWorker: resolve(DUCKDB_DIST, './duckdb-node-eh.worker.cjs')
+			}
+		};
+		const logger = new ConsoleLogger();
+
+		// and synchronous database
+		db = await createDuckDB(DUCKDB_BUNDLES, logger, NODE_RUNTIME);
+		await db.instantiate();
+		db.open({ query: { castBigIntToDouble: true, castTimestampToDate: true } });
+		connection = db.connect();
+		resolveInit();
+	} catch (e) {
+		rejectInit(e);
+		throw e;
+	}
 }
 
 /**
@@ -67,7 +86,7 @@ export async function setParquetURLs(urls) {
 		connection.query(`CREATE SCHEMA IF NOT EXISTS ${source};`);
 		for (const url of urls[source]) {
 			const table = url.split('/').at(-1).slice(0, -'.parquet'.length);
-			const file_name = `${table}.parquet`;
+			const file_name = `${source}_${table}.parquet`;
 			db.registerFileURL(file_name, `./static${url}`, DuckDBDataProtocol.NODE_FS, false);
 			connection.query(
 				`CREATE OR REPLACE VIEW ${source}.${table} AS SELECT * FROM read_parquet('${file_name}');`
@@ -80,14 +99,14 @@ export async function setParquetURLs(urls) {
  * Queries the database with the given SQL statement.
  *
  * @param {string} sql
- * @param {{ route_hash: string, query_name: string }} cache_options
+ * @param {{ route_hash: string, query_name: string, prerendering: boolean }} [cache_options]
  * @returns {import('apache-arrow').Table | null}
  */
 export function query(sql, cache_options) {
 	const res = connection.query(sql);
 
 	if (cache_options) {
-		cache_for_hash(cache_options.route_hash, sql, cache_options.query_name, res);
+		cache_for_hash(sql, res, cache_options);
 	}
 
 	return arrowTableToJSON(res);