+++ sites/example-project/src/pages/+layout.js
@@ -6,8 +6,9 @@ import {
 	query,
 	updateSearchPath
 } from '@evidence-dev/universal-sql/client-duckdb';
+import { profile } from '@evidence-dev/component-utilities/profile';
 
-const database_initialization = (async () => {
+const initDb = async () => {
 	let renderedFiles = {};
 
 	if (!browser) {
@@ -23,7 +24,9 @@ const database_initialization = (async () => {
 	await initDB();
 	await setParquetURLs(renderedFiles);
 	await updateSearchPath(Object.keys(renderedFiles));
-})();
+};
+
+const database_initialization = profile(initDb);
 
 /** @satisfies {import("./$types").LayoutLoad} */
 export const load = async ({
@@ -35,7 +38,7 @@ export const load = async ({
 	const data = {};
 
 	// let SSR saturate the cache first
-	if (!building && browser && isUserPage) {
+	if (browser && isUserPage) {
 		await Promise.all(
 			evidencemeta.queries?.map(async ({ id }) => {
 				const res = await fetch(`/api/${routeHash}/${id}.arrow`);
@@ -51,13 +54,7 @@ export const load = async ({
 					return database_initialization.then(() => query(sql));
 				}
 
-				const query_results = query(sql, { route_hash: routeHash, query_name });
-
-				// trigger the prerendering of the cache endpoint
-				// should make sure this isn't a race condition
-				fetch(`/api/${routeHash}/${query_name}.arrow`);
-
-				return query_results;
+				return query(sql, { route_hash: routeHash, query_name, prerendering: building });
 			}
 		},
 		data,